+++ spec/chef/recipes/gitlab-pages_spec.rb
@@ -31,6 +31,7 @@ describe 'gitlab::gitlab-pages' do
       expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-root-cert})
       expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-root-key})
       expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-metrics-address})
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-status-uri})
     end
   end
 
@@ -46,7 +47,8 @@ describe 'gitlab::gitlab-pages' do
           redirect_http: true,
           cert: '/etc/gitlab/pages.crt',
           artifacts_server_url: "https://gitlab.elsewhere.com/api/v5",
-          artifacts_server_timeout: 60
+          artifacts_server_timeout: 60,
+          status_uri: '/@status'
         }
       )
     end
@@ -68,6 +70,7 @@ describe 'gitlab::gitlab-pages' do
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-root-key})
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-artifacts-server="https://gitlab.elsewhere.com/api/v5"})
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-artifacts-server-timeout=60})
+      expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-pages-status="/@status"})
     end
   end